requirements.txt
azure-mgmt-compute
azure-identity
azure-mgmt-storage
azure-mgmt-resource
azure-mgmt-network
Flask
gevent
python-dotenv
wakeonlan